Add player.py
Move Player class and its subclasses to player.py
